refactor: [FE] 메인화면 레이아웃 구성
- 메인화면 레이아웃 설정(헤더, 사이드 메뉴, 오디오 파형)
- 전체적인 파일 네이밍 변경
- 본문 내용
