+++ frontend/public/index.html
@@ -8,7 +8,7 @@
 </head>
 
 <body>
-    <audio-editor-app id=root></audio-editor-app>
+    <audi-app id=root></audi-app>
 </body>
 
 </html>
