+++ frontend/src/components/AudioTrack/AudioTrack.scss
@@ -1,28 +1,5 @@
-*{
-    margin: 0;
-    padding: 0;
-}
-
-html,body{
-    width: 100%;
-    height: 100%;
-    margin: 0;
-    padding: 0;
-}
-
-body{
-    background-color: rgb(40,40,40);
-}
-
-#root{
-    width: 100%;
-    height: 100%;
-    margin: 0;
-    padding: 0;
-}
-
-.hide{
-    display: none;
+audio-track:not(:defined) {
+    display: hidden;
 }
 
 .audio-track-container{
