--- frontend/src/components/AudioTrack/index.js
@@ -1,1 +1,0 @@
-import "./AudioTrack.ts";
