+++ frontend/src/components/EditorMenu/EditorMenu.scss
