+++ frontend/src/components/Main/Main.scss
@@ -1,0 +1,20 @@
+.audi-main-container{
+    width: 100%;
+    height: 90%;
+    padding: 30px 20px;
+    box-sizing: border-box;
+    .audi-main-content{
+        display: flex;
+        justify-content: space-between;
+        width: 100%;
+        height: 100%;
+        padding: 30px;
+        border: 1px solid #ffffff;
+        border-radius: 5px;
+        background-color: #000000;
+
+        aside{
+            width: 20%;
+        }
+    }
+}
