+++ frontend/src/components/TimeInfo/TimeInfo.scss
