+++ frontend/src/components/index.js
@@ -1,9 +1,12 @@
-import './audio-editor-app';
-import './sample-component';
-import './AudioTrack';
-import './IconButton';
-import './Logo';
-import './EditorMenu';
-import './Header';
-import './EditTools';
-import './PlaybackTools';
+import './App/App';
+import './AudioTrack/AudioTrack';
+import './IconButton/IconButton';
+import './Logo/Logo';
+import './EditorMenu/EditorMenu';
+import './Header/Header';
+import './EditTools/EditTools';
+import './PlaybackTools/PlaybackTools';
+import './Sidebar/Sidebar';
+import './SourceList/SourceList';
+import './TimeInfo/TimeInfo';
+import './Main/Main';
