+++ frontend/src/components/sidebar/sidebar.scss
@@ -1,18 +1,26 @@
-#sidebar {
-    border:1px solid white;
-    background-color: black;
-    width:20%;
-    height:80%;
-    position:absolute;
-    padding:1px;
+side-bar:not(:defined) {
+    display: hidden;
+}
 
+#sidebar {
     display:flex;
     flex-direction:column;
     align-items:center;
+    padding:1px;
+    height:100%;
+
+    border:1px solid white;
+    border-radius: 5px;
+    background-color: black;
 }
 
 .sidebar-child {
     width:100%;
+}
+
+.sidebar-source-list {
     height:50%;
-    position:absolute;
+}
+.sidebar-time-info {
+    height:20%;
 }
