--- frontend/src/components/sidebar/style.scss
@@ -1,23 +1,0 @@
-#sidebar {
-    display:flex;
-    flex-direction:column;
-    align-items:center;
-    position:absolute;
-    padding:1px;
-    width:20%;
-    height:80%;
-
-    border:1px solid white;
-    background-color: black;
-}
-
-.sidebar-child {
-    width:100%;
-}
-
-.sidebar-source-list {
-    height:50%;
-}
-.sidebar-time-info {
-    height:20%;
-}
