--- frontend/src/components/source-list/source-list.scss
@@ -1,48 +1,0 @@
-.source-list-outer-wrap {
-    height:100%;
-    margin:5px;
-
-    display:flex;
-    flex-direction:column;
-    align-items: center;
-}
-
-.source-list-title {
-    width:98%;
-    height:10%;
-    div {
-        color:white;
-        margin:3px;
-    }
-}
-
-.source-list-wrap {
-    border:1px solid white;
-    height:80%;
-    width:95%;
-    margin-top:10px;
-}
-
-.audio-source {
-    background-color:black;
-    color:white;
-    margin:5px;
-    border:1px solid white;
-    height:30px;
-
-    &:hover {
-        background-color:rgb(60, 60, 60);
-        .source-info {
-            display:block;
-        };
-    }
-}
-
-.source-info {
-    display:none;
-    position:absolute;
-    background-color:grey;
-    width:60%;
-    left:30%;
-    height:60px;
-}
