+++ frontend/src/index.js
@@ -1,1 +1,2 @@
 import '@components';
+import '@style';
